Simplify enrichment handler
